componentized transactions-table
